Refactor static create methods into factories.
Tests can now be generic and operate on multiple concrete
implementations. I'd really like to make the factories
a generic type argument but I probably need to clean up
the interfaces a bit as the generic stuff is getting
a little crazy as it is.

All in all this makes it possible to have the correct type
intantiated by the factory but still have a common interface
for more borring operations.